fix(TCOMP-2049): move sanitize test to api
